getLong method should be using longValue

+++ impl/src/main/java/org/glassfish/json/JsonTokenizer.java
@@ -567,7 +567,7 @@ final class JsonTokenizer implements Closeable {
             }
             return minus ? -num : num;
         } else {
-            return getBigDecimal().intValue();
+            return getBigDecimal().longValue();
         }
     }
